fix(app) dqa software keyboard round3 (#13493)
* fix(app): dqa software keyboard round3
