+++ app/src/atoms/SoftwareKeyboard/CustomKeyboard/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Keyboard from 'react-simple-keyboard'
-import { customDisplay } from './constants'
+import { customDisplay } from '../constants'
 
 interface CustomKeyboardProps {
   onChange: (input: string) => void
@@ -17,7 +17,7 @@ const customLayout = {
   shift: [
     'Q W E R T Y U I O P',
     'A S D F G H J K L',
-    '{shift} Z X C V B N M {backspace}',
+    '{abc} Z X C V B N M {backspace}',
     '{numbers}',
   ],
   numbers: ['1 2 3', '4 5 6', '7 8 9', '{abc} 0 {backspace}'],
