+++ app/src/atoms/SoftwareKeyboard/Numpad/index.css
@@ -1,0 +1,7 @@
+/* stylelint-disable */
+
+.numpad button.hg-button.hg-button-backspace,
+.numpad button.hg-button.hg-button-abc,
+.numpad button.hg-button.hg-standardBtn {
+  flex: 1;
+}
