+++ app/src/atoms/SoftwareKeyboard/index.ts
@@ -1,3 +1,3 @@
+export { CustomKeyboard } from './CustomKeyboard'
 export { NormalKeyboard } from './NormalKeyboard'
 export { Numpad } from './Numpad'
-export { CustomKeyboard } from './CustomKeyboard'
